007, Add the reset game button and prep the event handler
When the user clicks the reset game button, the event handler currently
only console logs the three things that it needs to do, which are to
set new board, to reset the starting player to X, and to reset the
winner to null. Note that currently there are no state variables for
any of these three things.