013, Return a brand new board and update the state variable with it on each square click
To make updates to a state variable, you need to do it functionally, meaning, you
must clone the old value of a variable and save it as a new, updated variable - in this
case, the new variable is named newBoard.

Next, you'll invoke the setBoard state-updating function and pass it the newBoard as its
single argument.

This now makes it so that on each square button click, it's value is set to the letter X.